Link to the release app instead of Github
The current badger list shows two links for each entry.  The first links to the
repo of the application.  The second also links to Github but to the comparison
of `deployed-to-production` and `master`.

I think it would be a bit more useful to link directly to the release app so
you can get on with deploying the changes.

The only tricky thing is (I think) contacts-admin which is listed as `contacts`
in the release app (and govuk-puppet).

There were some more apps like this in the past, but I think they've all been
retired.  New apps shouldn't need any special handling as they should conform
to the [naming guidelines](https://docs.publishing.service.gov.uk/manual/naming.html#naming-applications).